% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{package}
\name{shinyscan-package}
\alias{shinyscan-package}
\alias{shinyscan}
\title{Shiny scan}
\description{
A collection of procedures for analyzing, visualizing, and managing
single-case data.
}
\author{
Juergen Wilbert [aut, cre]
}
\keyword{package}
